[compliance] feat(core): add server fetchers for DOT inspections and alerts (wave-1) (#38)
